requirements.txt
# SOP-gen Requirements
# AI-Powered SOP Document Generator

# Core Framework
streamlit>=1.28.0

# AI/LLM APIs
openai>=1.0.0
anthropic>=0.7.0

# Document Processing
python-docx>=0.8.11
markdown2>=2.4.10
xlsxwriter>=3.1.9

# PDF Generation (install one of these)
# Option 1: pdfkit (requires wkhtmltopdf system package)
pdfkit>=1.0.0
# Option 2: weasyprint (alternative, pure Python)
weasyprint>=60.0

# Data Handling
pandas>=2.0.0

# Utilities
python-dateutil>=2.8.2
python-dotenv>=1.0.0

# Phase 1 Enhancement: Translation and File Upload Support
deep-translator>=1.11.4
Pillow>=10.0.0
